+++ .env.example
@@ -8,3 +8,5 @@ DATABASE_URL=postgresql://${DATABASE_USER}:${DATABASE_PASSWORD}@db:${DATABASE_PO
 SECRET_KEY=
 ALGORITHM=
 ACCESS_TOKEN_EXPIRE_MINUTES=
+
+VITE_API_BASE_URL=http://localhost:8000
