+++ frontend/src/api/auth.ts
@@ -8,7 +8,8 @@ export async function signIn(email: string, password: string) {
   });
   if (!res.ok) {
     const error = await res.json().catch(() => ({}));
-    throw new Error(error?.message || "Failed to sign in. " + error?.detail);
+    console.error("Failed to sign in", error);
+    throw new Error("Failed to sign in. " + error?.detail);
   }
   return res.json();
 }
@@ -30,7 +31,8 @@ export async function signUp(
   });
   if (!res.ok) {
     const error = await res.json().catch(() => ({}));
-    throw new Error(error?.message || "Failed to sign up. " + error?.id);
+    console.error("Failed to sign up", error);
+    throw new Error("Failed to sign up. " + error?.id);
   }
   return res.json();
 }
