+++ frontend/src/components/SignIn.tsx
@@ -8,9 +8,7 @@ import {
 } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { useForm } from "react-hook-form";
-import { Link, useNavigate } from "react-router";
-
-import { signIn } from "@/api/auth";
+import { Link } from "react-router";
 import {
   Form,
   FormControl,
@@ -20,9 +18,8 @@ import {
   FormMessage,
 } from "@/components/ui/form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation } from "@tanstack/react-query";
 import { z } from "zod";
-import { toast } from "sonner";
+import { useSignIn } from "@/hooks/useSignIn";
 
 const formSchema = z.object({
   email: z.string().email(),
@@ -30,31 +27,13 @@ const formSchema = z.object({
 });
 
 export default function SignIn() {
+  const { mutate } = useSignIn();
+
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
       email: "",
       password: "",
-    },
-  });
-
-  const navigate = useNavigate();
-
-  const { mutate } = useMutation({
-    mutationFn: (data: { email: string; password: string }) =>
-      signIn(data.email, data.password),
-    onSuccess: (response) => {
-      toast.success("User logged in successfully", {
-        position: "top-right",
-      });
-
-      localStorage.setItem("access_token", response.access_token);
-      navigate("/");
-    },
-    onError: (error) => {
-      toast.error(error.message, {
-        position: "top-right",
-      });
     },
   });
