+++ frontend/src/hooks/useSignIn.tsx
@@ -1,0 +1,28 @@
+import { signIn } from "@/api/auth";
+import { useMutation } from "@tanstack/react-query";
+import { useNavigate } from "react-router";
+import { toast } from "sonner";
+
+export const useSignIn = () => {
+  const navigate = useNavigate();
+
+  const { mutate, isPending } = useMutation({
+    mutationFn: (data: { email: string; password: string }) =>
+      signIn(data.email, data.password),
+    onSuccess: (response) => {
+      toast.success("User logged in successfully", {
+        position: "top-right",
+      });
+
+      localStorage.setItem("access_token", response.access_token);
+      navigate("/");
+    },
+    onError: (error) => {
+      toast.error(error.message, {
+        position: "top-right",
+      });
+    },
+  });
+
+  return { mutate, isPending };
+};
